Don't attempt to kill process if no pid exists

+++ src/k8s-engine/steve.ts
@@ -69,6 +69,12 @@ export class Steve {
    * Stops the Steve API.
    */
   public stop() {
+    const { pid } = this.process || { };
+
+    if (!pid) {
+      return;
+    }
+
     this.process.kill();
   }
 }
